[功能修改] 在NPU中调用Mat和Vec

+++ include/cpu.h
@@ -5,8 +5,8 @@
 
 SC_MODULE(CPU) {
     sc_in<bool> clk;
-    sc_in<int> data_in;
-    sc_out<int> data_out;
+    sc_in<double> data_in;
+    sc_out<double> data_out;
     sc_out<bool> req, ack;
 
     void process();
